custom errors

+++ src/main/java/com/sanix/UserRegistrationSystem/Rest/UserRegistrationRestController.java
@@ -1,5 +1,6 @@
 package com.sanix.UserRegistrationSystem.Rest;
 
+import com.sanix.UserRegistrationSystem.Exception.CustomErrorType;
 import com.sanix.UserRegistrationSystem.dto.UserDTO;
 import com.sanix.UserRegistrationSystem.repository.UserJpaRepository;
 import com.sanix.UserRegistrationSystem.services.UserService;
@@ -23,27 +24,32 @@ public class UserRegistrationRestController {
             LoggerFactory.getLogger(UserRegistrationRestController.class);
 
     private UserService userService;
-    private UserJpaRepository userJpaRepository;
+    private final UserJpaRepository userJpaRepository;
 
-    public UserRegistrationRestController(UserJpaRepository userJpaRepository) {
+    public UserRegistrationRestController(UserService userService, UserJpaRepository userJpaRepository) {
+        this.userService = userService;
         this.userJpaRepository = userJpaRepository;
-    }
-
-    @Autowired
-    public void setUserJpaRepository(UserService userService)
-    {
-        this.userService=userService;
-
     }
 
     @GetMapping("/")
     public ResponseEntity<List<UserDTO>> listAllUsers(){
         List<UserDTO> users=userJpaRepository.findAll();
+        if(users.isEmpty()){
+            return new ResponseEntity<List<UserDTO>>(HttpStatus.NO_CONTENT);
+        }
+
         return new ResponseEntity<List<UserDTO>>(users, HttpStatus.OK);
     }
 
     @PostMapping(value="/", consumes= MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<UserDTO> createUser(@RequestBody final UserDTO user){
+        if(userService.findByUsername(user.getUsername()) != null){
+            return new ResponseEntity<UserDTO>(new CustomErrorType(
+                    "Unable to create new user. A user with name "
+                    +user.getUsername()+" already exist."
+            ), HttpStatus.CONFLICT);
+        }
+
         userJpaRepository.save(user);
         return new ResponseEntity<UserDTO>(user, HttpStatus.CREATED);
     }
@@ -52,6 +58,12 @@ public class UserRegistrationRestController {
     public ResponseEntity <UserDTO> getUserById(@PathVariable("id") final Long id){
 
         UserDTO user=userService.findById(id);
+        if(user==null){
+            return new ResponseEntity<UserDTO>(
+                    new CustomErrorType("User with id "
+                    +id+" not found"), HttpStatus.NOT_FOUND
+            );
+        }
 
         return new ResponseEntity <UserDTO>(user, HttpStatus.OK);
     }
@@ -62,8 +74,13 @@ public class UserRegistrationRestController {
         //fetch user based on id and set id to currentUSer object of type UserDTO
         UserDTO currentUser=userService.findById(id);
 
+        if(currentUser==null){
+            return new ResponseEntity<UserDTO>(new CustomErrorType("Unable to update. User with id " +
+                    id+" not found."), HttpStatus.NOT_FOUND);
+        }
+
         //update currentUser object data with user object data
-        currentUser.setName(user.getName());
+        currentUser.setUsername(user.getUsername());
         currentUser.setAddress(user.getAddress());
         currentUser.setEmail(user.getEmail());
 
@@ -77,6 +94,11 @@ public class UserRegistrationRestController {
     @DeleteMapping("/{id}")
     public ResponseEntity<UserDTO> deleteUser(@PathVariable("id") final Long id){
         UserDTO user=userService.findById(id);
+        if(user==null){
+            return new ResponseEntity<UserDTO>(
+                    new CustomErrorType("Unable to delete. User with id "
+                    +id+" not found ."), HttpStatus.NOT_FOUND);
+        }
         userJpaRepository.delete(user);
         return new ResponseEntity<UserDTO>(HttpStatus.NO_CONTENT);
     }
